requirements.txt
Flask==3.1.2
python-dotenv==1.0.1
prometheus-client==0.20.0
gunicorn==22.0.0
